+++ packages/core/src/services/documents/index.ts
@@ -4,6 +4,7 @@ export { default as deleteMultiple } from "./delete-multiple.js";
 export { default as deleteMultiplePermanently } from "./delete-multiple-permanently.js";
 export { default as deleteSingle } from "./delete-single.js";
 export { default as deleteSinglePermanently } from "./delete-single-permanently.js";
+export { default as duplicateSingle } from "./duplicate-single.js";
 export { default as getMultiple } from "./get-multiple.js";
 export { default as getMultipleRevisions } from "./get-multiple-revisions.js";
 export { default as getSingle } from "./get-single.js";